# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  barber: "barber",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  customer: "customer",
  barber: "barber",
} as const;

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type ShopStatus = (typeof ShopStatus)[keyof typeof ShopStatus];

export const ShopStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Shop {
  id: number;
  ownerId: number;
  name: string;
  description: string;
  address: string;
  city: string;
  phone: string;
  imageUrl?: string | null;
  status: ShopStatus;
  rejectionReason?: string | null;
  averageRating?: number | null;
  reviewCount: number;
  createdAt: string;
}

export interface Service {
  id: number;
  shopId: number;
  name: string;
  description?: string | null;
  price: number;
  durationMinutes: number;
}

export type ShopDetail = Shop & {
  services: Service[];
  ownerName: string;
};

export interface ShopListResponse {
  shops: Shop[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateShopRequest {
  name: string;
  description: string;
  address: string;
  city: string;
  phone: string;
  imageUrl?: string | null;
}

export interface UpdateShopRequest {
  name?: string;
  description?: string;
  address?: string;
  city?: string;
  phone?: string;
  imageUrl?: string | null;
}

export interface RejectShopRequest {
  reason: string;
}

export interface CreateServiceRequest {
  name: string;
  description?: string | null;
  price: number;
  durationMinutes: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  customerId: number;
  shopId: number;
  serviceId: number;
  status: BookingStatus;
  scheduledAt: string;
  notes?: string | null;
  shopName: string;
  serviceName: string;
  servicePrice: number;
  customerName: string;
  createdAt: string;
}

export interface CreateBookingRequest {
  shopId: number;
  serviceId: number;
  scheduledAt: string;
  notes?: string | null;
}

export type UpdateBookingStatusRequestStatus =
  (typeof UpdateBookingStatusRequestStatus)[keyof typeof UpdateBookingStatusRequestStatus];

export const UpdateBookingStatusRequestStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusRequest {
  status: UpdateBookingStatusRequestStatus;
}

export interface Review {
  id: number;
  shopId: number;
  customerId: number;
  customerName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}

export type ListShopsParams = {
  search?: string;
  city?: string;
  page?: number;
  limit?: number;
};

export type AdminListShopsParams = {
  status?: AdminListShopsStatus;
};

export type AdminListShopsStatus =
  (typeof AdminListShopsStatus)[keyof typeof AdminListShopsStatus];

export const AdminListShopsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;