# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Barber Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for current user
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (barber or customer)
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusRequest"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /shops/{shopId}/bookings:
    get:
      operationId: listShopBookings
      tags: [bookings]
      summary: List bookings for a shop (barber owner only)
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Shop bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /shops/{shopId}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/shops:
    get:
      operationId: adminListShops
      tags: [admin]
      summary: Admin - list all shops with status
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: All shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopListResponse"
  /admin/shops/{shopId}/approve:
    post:
      operationId: approveShop
      tags: [admin]
      summary: Admin - approve a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Shop approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /admin/shops/{shopId}/reject:
    post:
      operationId: rejectShop
      tags: [admin]
      summary: Admin - reject a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectShopRequest"
      responses:
        "200":
          description: Shop rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /admin/reviews/{reviewId}/moderate:
    delete:
      operationId: moderateReview
      tags: [admin]
      summary: Admin - remove a review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Review removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, barber, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [customer, barber]
      required:
        - email
        - password
        - name
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    Shop:
      type: object
      properties:
        id:
          type: number
        ownerId:
          type: number
        name:
          type: string
        description:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - name
        - description
        - address
        - city
        - phone
        - status
        - reviewCount
        - createdAt
...
    ShopListResponse:
      type: object
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - shops
        - total
        - page
        - limit
...
    RejectShopRequest:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...
    Booking:
      type: object
      properties:
        id:
          type: number
        customerId:
          type: number
        shopId:
          type: number
        serviceId:
          type: number
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        shopName:
          type: string
        serviceName:
          type: string
        servicePrice:
          type: number
        customerName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - shopId
        - serviceId
        - status
        - scheduledAt
        - shopName
        - serviceName
        - servicePrice
        - customerName
        - createdAt
    CreateBookingRequest:
      type: object
      properties:
        shopId:
          type: number
        serviceId:
          type: number
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required:
        - shopId
        - serviceId
        - scheduledAt
    UpdateBookingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, completed, cancelled]
      required:
        - status
    Review:
      type: object
      properties:
        id:
          type: number
        shopId:
          type: number
        customerId:
          type: number
        customerName:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - shopId
        - customerId
        - customerName
        - rating
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
      required:
        - rating