# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  customerId: number;
  shopId: number;
  serviceId: number;
  status: BookingStatus;
  scheduledAt: Date;
  notes?: string | null;
  shopName: string;
  serviceName: string;
  servicePrice: number;
  customerName: string;
  createdAt: Date;
}