# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  shopId: number;
  serviceId: number;
  scheduledAt: Date;
  notes?: string | null;
}